Fix whitespace detection and restructure Token type
In Java whitespace is defined as one of a handful of chars and
not in terms of the unicode property whitespace. Fixed that.

The `Token` type needed some restructuring to match the
specification more closely. Comments explaining where it does not
match were also added. Some token variants were removed or
renamed.